Feat/transactions (#8)
* Added err handling on update user

* Updated SQL query build. Added result verification. Updated migrations.

* Added from/to to get all transactions

* Added get category by ID. Updated categories migration.

* Updated ids from int to int64. Updated integer to bigint inmigrations. Removed state from users table.

* Added compose. Updated Dockerfile. Refactor postgres URI. Added envs. Updated .env.example.

* Updated to latest version of alpine

+++ internal/controller/users.go
@@ -33,7 +33,7 @@ func (c *userController) Create(ctx *gin.Context) {
 		return
 	}
 
-	body.ID = ctx.GetInt(defines.ParamUserID)
+	body.ID = ctx.GetInt64(defines.ParamUserID)
 
 	if !body.IsValid() {
 		ctx.JSON(http.StatusBadRequest, defines.ErrInvalidBody)
@@ -50,7 +50,7 @@ func (c *userController) Create(ctx *gin.Context) {
 	ctx.JSON(http.StatusCreated, jsend.NewSuccess(response))
 }
 func (c *userController) Get(ctx *gin.Context) {
-	userID := ctx.GetInt(defines.ParamUserID)
+	userID := ctx.GetInt64(defines.ParamUserID)
 
 	response, err := c.srv.GetByID(userID)
 
@@ -69,7 +69,7 @@ func (c *userController) Update(ctx *gin.Context) {
 		return
 	}
 
-	body.ID = ctx.GetInt(defines.ParamUserID)
+	body.ID = ctx.GetInt64(defines.ParamUserID)
 
 	if !body.IsValid() {
 		ctx.JSON(http.StatusBadRequest, defines.ErrInvalidBody)
@@ -86,7 +86,7 @@ func (c *userController) Update(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, jsend.NewSuccess(response))
 }
 func (c *userController) Delete(ctx *gin.Context) {
-	userID := ctx.GetInt(defines.ParamUserID)
+	userID := ctx.GetInt64(defines.ParamUserID)
 
 	err := c.srv.Delete(userID)
