+++ internal/service/wallets.go
@@ -8,11 +8,11 @@ import (
 
 type WalletsService interface {
 	Create(walletDTO *domain.WalletDTO) (*domain.WalletDTO, error)
-	GetByName(userID int, name string) (*domain.WalletDTO, error)
-	GetByID(userID int, id int) (*domain.WalletDTO, error)
-	GetAll(userID int) (*[]domain.WalletDTO, error)
+	GetByName(userID int64, name string) (*domain.WalletDTO, error)
+	GetByID(userID int64, id int64) (*domain.WalletDTO, error)
+	GetAll(userID int64) (*[]domain.WalletDTO, error)
 	UpdateByID(walletDTO *domain.WalletDTO) (*domain.WalletDTO, error)
-	DeleteByID(id int, userID int) error
+	DeleteByID(id int64, userID int64) error
 }
 type walletsService struct {
 	repo repository.WalletRepository
@@ -35,7 +35,7 @@ func (s *walletsService) Create(walletDTO *domain.WalletDTO) (*domain.WalletDTO,
 
 	return wallet.ToDTO(), nil
 }
-func (s *walletsService) GetByName(userID int, name string) (*domain.WalletDTO, error) {
+func (s *walletsService) GetByName(userID int64, name string) (*domain.WalletDTO, error) {
 	wallet, err := s.repo.GetBySanitizedName(sanitizer.Sanitize(name), userID)
 	if err != nil {
 		return nil, err
@@ -43,7 +43,7 @@ func (s *walletsService) GetByName(userID int, name string) (*domain.WalletDTO, 
 
 	return wallet.ToDTO(), nil
 }
-func (s *walletsService) GetByID(userID int, id int) (*domain.WalletDTO, error) {
+func (s *walletsService) GetByID(userID int64, id int64) (*domain.WalletDTO, error) {
 	wallet, err := s.repo.GetByID(id, userID)
 	if err != nil {
 		return nil, err
@@ -51,7 +51,7 @@ func (s *walletsService) GetByID(userID int, id int) (*domain.WalletDTO, error) 
 
 	return wallet.ToDTO(), nil
 }
-func (s *walletsService) GetAll(userID int) (*[]domain.WalletDTO, error) {
+func (s *walletsService) GetAll(userID int64) (*[]domain.WalletDTO, error) {
 	var err error
 	var wallets *[]domain.Wallet
 
@@ -79,6 +79,6 @@ func (s *walletsService) UpdateByID(walletDTO *domain.WalletDTO) (*domain.Wallet
 
 	return wallet.ToDTO(), nil
 }
-func (s *walletsService) DeleteByID(id int, userID int) error {
+func (s *walletsService) DeleteByID(id int64, userID int64) error {
 	return s.repo.DeleteByID(id, userID)
 }
